+++ platform/default/mbgl/storage/offline_database.cpp
@@ -51,7 +51,7 @@ void OfflineDatabase::ensureSchema() {
             case 0: break; // cache-only database; ok to delete
             case 1: break; // cache-only database; ok to delete
             case 2: migrateToVersion3(); // fall through
-            case 3: indexURLTemplates(); return;
+            case 3: checkURLTemplateIndexing(); return;
             default: throw std::runtime_error("unknown schema version");
             }
 
@@ -301,7 +301,14 @@ bool OfflineDatabase::putResource(const Resource& resource,
 }
 
 optional<std::pair<Response, uint64_t>> OfflineDatabase::getTile(const Resource::TileData& tile) {
-    Statement accessedStmt = getStatement(
+    Statement accessedStmt = getStatement(nonIndexedURLTemplates ?
+        "UPDATE tiles "
+        "SET accessed       = ?1 "
+        "WHERE url_template = ?2 "
+        "  AND pixel_ratio  = ?3 "
+        "  AND x            = ?4 "
+        "  AND y            = ?5 "
+        "  AND z            = ?6 ":
         "UPDATE tiles "
         "SET accessed           = ?1 "
         "WHERE url_template_id  = ( "
@@ -321,7 +328,15 @@ optional<std::pair<Response, uint64_t>> OfflineDatabase::getTile(const Resource:
     accessedStmt->bind(6, tile.z);
     accessedStmt->run();
 
-    Statement stmt = getStatement(
+    Statement stmt = getStatement(nonIndexedURLTemplates ?
+        //        0      1        2       3        4
+        "SELECT etag, expires, modified, data, compressed "
+        "FROM tiles "
+        "WHERE url_template = ?1 "
+        "  AND pixel_ratio  = ?2 "
+        "  AND x            = ?3 "
+        "  AND y            = ?4 "
+        "  AND z            = ?5 ":
         //        0      1        2       3        4
         "SELECT etag, expires, modified, data, compressed "
         "FROM tiles "
@@ -369,7 +384,15 @@ bool OfflineDatabase::putTile(const Resource::TileData& tile,
                               const std::string& data,
                               bool compressed) {
     if (response.notModified) {
-        Statement update = getStatement(
+        Statement update = getStatement(nonIndexedURLTemplates ?
+            "UPDATE tiles "
+            "SET accessed       = ?1, "
+            "    expires        = ?2 "
+            "WHERE url_template = ?3 "
+            "  AND pixel_ratio  = ?4 "
+            "  AND x            = ?5 "
+            "  AND y            = ?6 "
+            "  AND z            = ?7 ":
             "UPDATE tiles "
             "SET accessed           = ?1, "
             "    expires            = ?2 "
@@ -399,7 +422,19 @@ bool OfflineDatabase::putTile(const Resource::TileData& tile,
     // to INSERT a resource at the same moment.
     Transaction transaction(*db, Transaction::Immediate);
 
-    Statement update = getStatement(
+    Statement update = getStatement(nonIndexedURLTemplates ?
+        "UPDATE tiles "
+        "SET modified       = ?1, "
+        "    etag           = ?2, "
+        "    expires        = ?3, "
+        "    accessed       = ?4, "
+        "    data           = ?5, "
+        "    compressed     = ?6 "
+        "WHERE url_template = ?7 "
+        "  AND pixel_ratio  = ?8 "
+        "  AND x            = ?9 "
+        "  AND y            = ?10 "
+        "  AND z            = ?11 ":
         "UPDATE tiles "
         "SET modified           = ?1, "
         "    etag               = ?2, "
@@ -440,21 +475,19 @@ bool OfflineDatabase::putTile(const Resource::TileData& tile,
         return false;
     }
 
-    Statement insert1 = getStatement(
-        "INSERT OR IGNORE INTO url_templates (url_template) "
-        "VALUES                              (?1) ");
-    
-    insert1->bind(1, tile.urlTemplate);
-    
-    insert1->run();
+    if (! nonIndexedURLTemplates) {
+        Statement insert = getStatement(
+            "INSERT OR IGNORE INTO url_templates (url_template) "
+            "VALUES                              (?1) ");
+        
+        insert->bind(1, tile.urlTemplate);
+        
+        insert->run();
+    }
 
-    Statement insert2 = nonIndexedURLTemplates ? getStatement(
-        "INSERT INTO tiles (url_template, url_template_id, pixel_ratio, x,  y,  z,  modified,  etag,  expires,  accessed,  data, compressed) "
-        "VALUES            (?1,           ( "
-        "                   SELECT id "
-        "                   FROM url_templates "
-        "                   WHERE url_template = ?1 ), "
-        "                                                  ?2,          ?3, ?4, ?5, ?6,        ?7,    ?8,       ?9,        ?10,  ?11) "): getStatement(
+    Statement insert = getStatement(nonIndexedURLTemplates ?
+        "INSERT INTO tiles (url_template, pixel_ratio, x,  y,  z,  modified,  etag,  expires,  accessed,  data, compressed) "
+        "VALUES            (?1,           ?2,          ?3, ?4, ?5, ?6,        ?7,    ?8,       ?9,        ?10,  ?11) ":
         "INSERT INTO tiles (url_template_id, pixel_ratio, x,  y,  z,  modified,  etag,  expires,  accessed,  data, compressed) "
         "VALUES            (( "
         "                   SELECT id "
@@ -462,25 +495,25 @@ bool OfflineDatabase::putTile(const Resource::TileData& tile,
         "                   WHERE url_template = ?1 ), "
         "                                    ?2,          ?3, ?4, ?5, ?6,        ?7,    ?8,       ?9,        ?10,  ?11) ");
 
-    insert2->bind(1, tile.urlTemplate);
-    insert2->bind(2, tile.pixelRatio);
-    insert2->bind(3, tile.x);
-    insert2->bind(4, tile.y);
-    insert2->bind(5, tile.z);
-    insert2->bind(6, response.modified);
-    insert2->bind(7, response.etag);
-    insert2->bind(8, response.expires);
-    insert2->bind(9, util::now());
+    insert->bind(1, tile.urlTemplate);
+    insert->bind(2, tile.pixelRatio);
+    insert->bind(3, tile.x);
+    insert->bind(4, tile.y);
+    insert->bind(5, tile.z);
+    insert->bind(6, response.modified);
+    insert->bind(7, response.etag);
+    insert->bind(8, response.expires);
+    insert->bind(9, util::now());
 
     if (response.noContent) {
-        insert2->bind(10, nullptr);
-        insert2->bind(11, false);
+        insert->bind(10, nullptr);
+        insert->bind(11, false);
     } else {
-        insert2->bindBlob(10, data.data(), data.size(), false);
-        insert2->bind(11, compressed);
+        insert->bindBlob(10, data.data(), data.size(), false);
+        insert->bind(11, compressed);
     }
 
-    insert2->run();
+    insert->run();
     transaction.commit();
 
     return true;
@@ -555,7 +588,15 @@ uint64_t OfflineDatabase::putRegionResource(int64_t regionID, const Resource& re
 
 bool OfflineDatabase::markUsed(int64_t regionID, const Resource& resource) {
     if (resource.kind == Resource::Kind::Tile) {
-        Statement insert = getStatement(
+        Statement insert = getStatement(nonIndexedURLTemplates ?
+            "INSERT OR IGNORE INTO region_tiles (region_id, tile_id) "
+            "SELECT                              ?1,        tiles.id "
+            "FROM tiles "
+            "WHERE url_template = ?2 "
+            "  AND pixel_ratio  = ?3 "
+            "  AND x            = ?4 "
+            "  AND y            = ?5 "
+            "  AND z            = ?6 ":
             "INSERT OR IGNORE INTO region_tiles (region_id, tile_id) "
             "SELECT                              ?1,        tiles.id "
             "FROM tiles "
@@ -580,7 +621,16 @@ bool OfflineDatabase::markUsed(int64_t regionID, const Resource& resource) {
             return false;
         }
 
-        Statement select = getStatement(
+        Statement select = getStatement(nonIndexedURLTemplates ?
+            "SELECT region_id "
+            "FROM region_tiles, tiles "
+            "WHERE region_id   != ?1 "
+            "  AND url_template = ?2 "
+            "  AND pixel_ratio  = ?3 "
+            "  AND x            = ?4 "
+            "  AND y            = ?5 "
+            "  AND z            = ?6 "
+            "LIMIT 1 ":
             "SELECT region_id "
             "FROM region_tiles, tiles "
             "INNER JOIN url_templates "
@@ -761,7 +811,11 @@ uint64_t OfflineDatabase::getOfflineMapboxTileCount() {
         return *offlineMapboxTileCount;
     }
 
-    Statement stmt = getStatement(
+    Statement stmt = getStatement(nonIndexedURLTemplates ?
+        "SELECT COUNT(DISTINCT id) "
+        "FROM region_tiles, tiles "
+        "WHERE tile_id = tiles.id "
+        "AND url_template LIKE 'mapbox://%' ":
         "SELECT COUNT(DISTINCT tiles.id) "
         "FROM region_tiles, tiles "
         "INNER JOIN url_templates "
@@ -775,43 +829,28 @@ uint64_t OfflineDatabase::getOfflineMapboxTileCount() {
     return *offlineMapboxTileCount;
 }
     
-bool OfflineDatabase::indexURLTemplates()
+void OfflineDatabase::checkURLTemplateIndexing()
 {
     Statement checkTiles = getStatement("PRAGMA table_info(tiles)");
     Statement checkURLTemplates = getStatement("PRAGMA table_info(url_templates)");
-    bool urlTemplatesMissing = ! checkURLTemplates->run();
+    bool urlTemplatesPresent = checkURLTemplates->run();
+    bool urlTemplatePresent = false;
+    bool urlTemplateIDPresent = false;
 
-    while (checkTiles->run() && ! nonIndexedURLTemplates) {
+    while (checkTiles->run()) {
         std::string columnName = checkTiles->get<std::string>(1);
         if (columnName == "url_template") {
-            nonIndexedURLTemplates = true;
-        }
-    }
-
-    if (urlTemplatesMissing) {
-        try {
-            Transaction transaction(*db, Transaction::Immediate);
-            db->exec("CREATE TABLE url_templates (\n"
-                     "  id INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,\n"
-                     "  url_template TEXT NOT NULL,\n"
-                     "  UNIQUE (url_template)\n"
-                     ");\n"
-                     "INSERT INTO url_templates (url_template) "
-                     "  SELECT DISTINCT url_template FROM tiles;\n"
-                     "ALTER TABLE tiles "
-                     "  ADD COLUMN url_template_id INTEGER REFERENCES url_templates(id);\n"
-                     "UPDATE tiles SET url_template_id = ( "
-                     "    SELECT id "
-                     "    FROM url_templates "
-                     "    WHERE url_template = tiles.url_template );\n");
-            transaction.commit();
-        } catch (...) {
-            Log::Error(Event::Database, "Unexpected error updating database schema: %s", util::toString(std::current_exception()).c_str());
-            throw;
+            urlTemplatePresent = true;
+        } else if (columnName == "url_template_id") {
+            urlTemplateIDPresent = true;
         }
     }
     
-    return urlTemplatesMissing;
+    nonIndexedURLTemplates = urlTemplatePresent && ! urlTemplateIDPresent;
+    
+    if (nonIndexedURLTemplates ? urlTemplatesPresent: ! (urlTemplatesPresent && ! urlTemplatePresent && urlTemplateIDPresent)) {
+        Log::Warning(Event::Database, "Unexpected result from database schema inspection");
+    }
 }
     
 } // namespace mbgl